GeoEnergyIO

Documentation for GeoEnergyIO.

Parsing of simulation cases

GeoEnergyIO.InputParser.parse_data_file — Function

parse_data_file(filename; unit = :si)
data = parse_data_file("MY_MODEL.DATA")

Parse a .DATA file given by the filename (industry standard input file) into a Dict. Units will be converted to strict SI unless you pass something else like units = :field. Setting units = nothing will skip unit conversion. Note that JutulDarcy assumes that the unit system is internally consistent. It is highly recommended to parse to the SI units if you want to perform simulations.

The best publicly available documentation on this format is available from the Open Porous Media (OPM) project's webpages: OPM Flow manual .

Keyword arguments

warn_parsing=true: Produce a warning when keywords are not supported (or partially supported) by the parser.

warn_feature=true`: Produce a warning when keywords are supported, but have limited or missing support in the numerical solvers.

units=:si: Symbol that indicates the unit system to be used in the output. Setting this to nothing will return values without conversion, i.e. exactly what is in the input files. :si will use strict SI. Other alternatives are :field and :metric. :lab is currently unsupported.

Note

This function is experimental and only covers a small portion of the keywords that exist for various simulators. You will get warnings that indicate the level of support for keywords in both the parser and the numerical solvers when known keywords with limited support. Pull requests for new keywords are welcome!

Parsing and processing of corner-point grids

GeoEnergyIO.InputParser.parse_grdecl_file — Function

parse_grdecl_file("mygrid.grdecl"; actnum_path = missing, kwarg...)

Parse a GRDECL file separately from the full input file. Note that the GRID section does not contain units - passing the input_units keyword is therefore highly recommended.

Keyword arguments

units=:si: Units to use for return values. Requires input_units to be set.

input_units=nothing: The units the file is given in.

verbose=false: Toggle verbosity.

GeoEnergyIO.CornerPointGrid.mesh_from_grid_section — Function

mesh_from_grid_section(f, actnum = missing)

TBW

Utilities

GeoEnergyIO.InputParser.get_data_file_cell_region — Function

get_data_file_cell_region(outer_data, t::Symbol; active = nothing)

Get the region indicator of some type for each cell of the domain.

t should be one of :satnum, :pvtnum or :eqlnum.

GeoEnergyIO.InputParser.number_of_tables — Function

number_of_tables(outer_data, t::Symbol)

Number of tables for given type t: one of :satnum, :pvtnum or :eqlnum.

GeoEnergyIO.CornerPointGrid.mesh_from_grid_section

GeoEnergyIO.InputParser.get_data_file_cell_region

GeoEnergyIO.InputParser.number_of_tables

GeoEnergyIO.InputParser.parse_data_file

GeoEnergyIO.InputParser.parse_grdecl_file