GeoEnergyIO

Documentation for GeoEnergyIO. See the subpages on the left for more details on functionality.

Internals

GeoEnergyIO.InputParser.keyword_default_value — Function

keyword_default_value(x::AbstractString, T::Type)

Get the default value of a keyword (as String or ::Val{X} where X is a Symbol) when placed in a array with element type T. This is used to initialize defaulted entries when using COPY, ADD, MULTIPLY and so on.

Index of functions

GeoEnergyIO.CornerPointGrid.cpgrid_from_horizons

GeoEnergyIO.CornerPointGrid.mesh_from_grid_section

GeoEnergyIO.InputParser.get_data_file_cell_region

GeoEnergyIO.InputParser.keyword_default_value

GeoEnergyIO.InputParser.number_of_tables

GeoEnergyIO.InputParser.parse_data_file

GeoEnergyIO.InputParser.parse_grdecl_file

GeoEnergyIO.InputParser.skip_kw!

GeoEnergyIO.read_egrid

GeoEnergyIO.read_init

GeoEnergyIO.read_restart

GeoEnergyIO.read_summary

GeoEnergyIO.write_egrid

GeoEnergyIO.write_jutuldarcy_summary